requirements.txt
flask>=3.0
anthropic>=0.42
python-dotenv>=1.0
pytest>=8.0
